[core][mem][cma] calloc needs to zero the memory
This is potentially a performance problem because the underlying
allocator might be able to do this faster (e.g., if it knows its
mmap(2)'ing pages which are zeroed upon initialization). But we
can't easily pass through even by creating a new calloc primitive,
because we can't use calloc(3) even for malloc allocator: we need
aligned_alloc.