[execution][blockcap] add blockcap library
The block capture ("blockcap") library extends the generic event capture
("evcap") system with utilities for writing captures in a special form
when they are known to contain execution events. A "blockcap" file is
just a regular event capture file except that it is expected to have
particular sections in it related to finalized blocks.

There are four related utilities included in this commit:

  - The blockcap builder collects all execution events in a
    [BLOCK_START, BLOCK_END] range into a "proposal" object, the core
    of which is a vbuf that records all events in the block's scope

  - The blockcap finalize tracker observes consensus events and
    computes which proposals are finalized or abandoned

  - The blockcap pack writer creates files containing multiple event
    sections, one for each finalized proposal. It also writes a block
    index

  - The blockcap archive is used as an API for reading and writing the
    local block archive