[event] move Linux-dependent functions to their own file
Most of the event ring library is simple C23 and atomic operations, a
few parts require Linux-specific behavior. Although the library isn't
supported anywhere else, making it compile as if it was helps some
Rust workflows, e.g., building the cargo docs locally on macOS will
need to compile it.