[event] export the event ring recorder in the SDK
Rust is going to use this to record its own consensus events. Along with
it, also include align.h and the transitively-included assert.h in the
libmonad_event.a build.

If the end-user ends up calling recorder functions, they'll need to provide
a `monad_assertion_failed` symbol to the linker, since this is not
otherwise part of libmonad_event.a.

When used by libmonad_execution, this symbol (from `assert.c`) pulls in
libboost_stacktrace.so and the entire C++ runtime. Whomever is writing the
final application should have some defined termination behavior this can
hook into.