[event] refactor the event ring and its iterator
This commit is motivated by the addition of the event capture file
format and API, which is added in a later commit.

The event capture reader also has an iterator, and it has a similar
API as the event ring iterator. Later on, a third API is added that
is agnostic between the two: it works with event rings, capture files,
or a type-erased "any source" type.

This commit renames various things so that all three APIs look more
similar. It moves most of the heavy lifting out of the iterator and
into the event ring. It also moves some defintions shared by event
rings and event captures into `event_def.h` so that the latter will
not need to include `event_ring.h`.

Finally, it cleans up a long-standing confusing implementation detail
where the iterator stored the "last read sequence number". The "current"
sequence number (the next to be consumed) is indexed by this, because
of the correspondence that the array index of the next event is the
same as the last read sequence number. This was very confusing and error
prone to reason about. Now we store the current sequence number, which
must be at least 1.