[event] add the event source C and C++ APIs
This commit adds an API that generalizes reading from either an event
ring or an event capture file. It offers the functionality of both APIs
for either type using static polymorphism, using _Generic in C and
function overloading in C++.

It also adds a type erased event source, for use in code that does
not care what the underlying type is.