[call-tracer] CallFrame::Log stores receipt_index
There two changes here:

- State::store_log returns the index in the receipt's log array that
  the log will have, and it becomes [[nodiscard]] because whomever is
  calling it shouldn't forget to propagate it to the call tracer

- CallTracer::on_log is passed the receipt index, and records it into
  the CallFrame

The rationale is to support the call frame<->log association in the
execution events.

It does not make sense to record the logs twice, especially if they
are large. They will instead be recorded by their receipt index,
which is the same as the event recording order.