[async] Refactor chunk type
 * Rename chunk -> chunk_t
 * Rename device -> device_t

These are shadowed by methods with the same name.

 * Remove cnv_chunk and seq_chunk

These types serve no purpose. When the chunk_t type is constructed, the
append_only constructor argument already does the same thing as the
method overrides. As far as I can tell, these are only for declarative
purposes. But removing the inheritance will set us up to use a vector
instead of shared_ptr.