[sdk-only] limited macOS OS compatibility
This allows the example program to work with snapshots, so that the SDK
can be used on a local development machine. This declares and defines
only enough of the C23 (and Linux) functions that may be missing from
libc (or the OS) so that the reader APIs work. We do not care about
supporting the writer APIs.