[core][event] add a C++ event recording interface
This factors out the C++ event recording infrastructure that is part
of exec_event_recorder.{hpp,cpp}. In a subsequent commit, it will
be changed to these utility class.

The reason for doing this is that currently, there is only a single
recorder, for the execution event ring. In later commits, other
recorders exist which have their own C++ helper objects. This moves
the common functionality to a new file.