[event] refactor exclusive writer detection
Event rings are unlinked from the filesystem upon graceful shutdown, and
re-created when a process restarts. If a process exits in an ungraceful
way, e.g. SIGKILL, it may leave a zombie file around. Most "ring death
detection" strategies reduce to detecting when writer processes exit,
which can always be done reliably, typically with pidfd_open(2) file
descriptors on Linux.

Previously, the reader would discover the association of a ring with
writer processes by scraping the fcntl.h "open flags" from the file
descriptor table in the proc(5) filesystem. Whomever had the ring open
with O_RDWR or O_WRONLY was considered a writer. This was done with a
helper function called `monad_event_ring_find_writer_pids`.

A simpler, faster, and less brittle way is to use flock(2) to communicate
the same information. flock(2) is used anyway to prevent writer processes
from re-initializing a non-zombie ring when a second daemon has been
started in error.

It is easier to scrape `/proc/locks` than scraping the file descriptor
table of every accessible process. Also, the lock table is deliberately
public, so we don't have to worry about credential differences between
reader and writer wrt/ the reader's ability to read the writer's proc(5)
entries.

The scheme followed here is that "exclusive writers" (single writer
process for a ring) acquire exclusive locks (LOCK_EX), and a potential
"multi-process writer" scenario could use shared locks (LOCK_SH).
Currently we only need the exclusive writer case.

The code also moved to a special Linux-specific file. Although the
execution daemon cannot be run on non-Linux hosts, the event library can
compile on other systems. This helps with a few workflows:

- Certain Rust workflows, e.g., building the cargo docs locally on macOS
  to look at them in a brower, runs enough of the build system that it
  tries to compile the monad-event-ring package. This in turn runs the
  compiler, so it's useful for the SDK to be able to compile in more
  places

- Several developers have reached out about trying the SDK example
  program on a local macOS development laptop; in response, the first
  two steps of the tutorial were changed to explicitly allow this, to
  lower the barrier to entry for trying out the SDK

The flock-based approach is also easier to implement on any UNIX
variant. flock information is _meant_ to be read, so almost all
systems have an easy way to read it.