[trace] Native transfer logs
This patch implements the execution client plumbing for the
[`traceTransfers` flag in
`eth_simulatev1`](https://geth.ethereum.org/docs/interacting-with-geth/rpc/ns-eth).This
flag causes ERC-20 style log events to get emitted when native token
transfers are made (on calls, contract creation, and selfdestruction).

The flag is not yet wired up to anything except some unit tests. When the
rest of `eth_simulatev1` is implemented, enabling the feature is a
straightforward integration.