[execution/ethereum]: Remove validate_output_header from chain
Having this method on the Chain object is holdover from a time when the
replay ethereum and live execution had the same entrypoint.

With the runloops, this is no longer necessary.

Header validation on the chain is a bit of a footgun. Replaying monad
from ethereum blocks uses the Monad chain with only live execution
validation. Replay monad now checks the state root.