[event] add monad_event_ring_wait_for_excl_writer
This API function waits for the event ring file to be created, if
it does not already exist. It also checks that the event ring file
has an exclusive writer process associated with it, and optionally
returns the pid of that writer process. It can also optionally return
a file descriptor for the file, with the given open(2) flags (this new
function must call `monad_event_ring_query_excl_writer_pid` internally,
which has to open the file anyway).

Rationale for this function:

When a reader process connects to an event ring that has an exclusive
writer (e.g., an active execution daemon, for execution events) it
has to decide what to do when that process terminates (when it
restarts, crashes, etc.).

There are two models:

  1. The reader process can die too

  2. The reader process can wait for the writer process to come back

The choice is up to program author, but the SDK allows both models.
In the next few commits we'll introduce the event-hist-recovery
example program and the blockcapd daemon, which both use the second
model.