[core] rename get_tl_tid() -> monad_thread_get_id()
The main thing is to make the name more consistent with the rest of
the C code (e.g, monad_statesync_, monad_executor_, monad_event, etc.)

While here, this also introduces a typedef `monad_tid_t` instead of
directly using pid_t. monad_tid_t is also long and not int (like pid_t)
so it changes the size of one of the classes

The reason for doing this is that system thread IDs are long and not
pid_t in most other systems. It helps the event SDK compile on macOS,
where parts of it might be used for "offline" data analysis or research