[ethereum/db]: Add CommitBuilder API
TrieDb::commit() has become a large function that will continue to grow
as new features are added. The CommitBuilder API contains the logic for
building a coherent UpdateList to TrieDb, and commit becomes a small
shim around db.upsert().

This is the first step in a larger refactor. Ideally, we should be
passing an UpdateList into commit(), and the caller should use the
CommitBuilder API to construct a set of Updates without adding many
lines of code.

This has two benefits:

* Test cases that only commit a subset of the tables don't need to
   build every field.

* This lends well to future upgrades where Monad will diverge from
  Ethereum. The caller can construct the relevant set of updates for a
  chain.